Add permission checks to backed

+++ app/Http/Controllers/Admin/UserController.php
@@ -7,6 +7,7 @@ use Inertia\Inertia;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\UserRequest;
+use Illuminate\Support\Facades\Auth;
 use Spatie\Permission\Models\Role;
 
 class UserController extends Controller
@@ -16,6 +17,10 @@ class UserController extends Controller
      */
     public function index(Request $request)
     {
+        if (! Auth::user()->canAny(['view_users', 'create_users', 'edit'])) {
+            abort(403);
+        }
+
         $users = User::with('roles:name')->latest()->paginate($request->per_page ?? 10);
         $roles = $roles = Role::where('panel', 'admin')->get()->pluck('name');
         return Inertia::render("Admin/Users/Index", compact('users', 'roles'));
@@ -26,6 +31,10 @@ class UserController extends Controller
      */
     public function store(UserRequest $request)
     {
+        if (! Auth::user()->can('create_users')) {
+            abort(403);
+        }
+
         $input = $request->validated();
         $role = $input['role'];
 
@@ -51,6 +60,10 @@ class UserController extends Controller
      */
     public function update(UserRequest $request, User $user)
     {
+        if (! Auth::user()->can('edit_users')) {
+            abort(403);
+        }
+
         $input = $request->validated();
         $role = $input['role'];
 
@@ -68,6 +81,10 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
+        if (! Auth::user()->can('delete_users')) {
+            abort(403);
+        }
+
         $user->delete();
 
         return back();
@@ -75,6 +92,10 @@ class UserController extends Controller
 
     public function destroyMany(Request $request)
     {
+        if (! Auth::user()->can('delete_users')) {
+            abort(403);
+        }
+        
         $request->validate([
             'ids' => 'required|array'
         ]);
